requirements.txt
pandas
elasticsearch
requests
python-dotenv
streamlit